requirements.txt
google-auth
google-auth-oauthlib 
google-api-python-client 
langchain-core 
# langchain-openai
streamlit 
pandas
python-dotenv
langchain-google-genai